# back-ogs/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OGS",
        "contact": {
            "email": "mickpicart@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "OGS API"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "JWT"
                ],
                "summary": "Create a JWT via given credentials",
                "description": "Returns the JWT",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully connected and token created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/auth/register": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Register a new user",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The user firstname and lastname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email that will be used for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password that will be used for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "To confirm the password that will be used for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully registered",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "JWT"
                ],
                "summary": "Sign out user",
                "description": "Revoke JWT",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "User successfully signed out",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/auth/user-profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get the authenticated user",
                "description": "Returns the authenticated user",
                "operationId": "userProfile",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/auth/password/reset-link": {
            "put": {
                "tags": [
                    "Password"
                ],
                "summary": "Send an email containing the password reset link with a 10 min token",
                "description": "Returns password reset email sent status",
                "operationId": "resetLink",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email where reset link will be sent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/auth/password/reset": {
            "patch": {
                "tags": [
                    "Password"
                ],
                "summary": "Modify user password in DB",
                "description": "Returns password reset confirmation",
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "The token catched from the url in the email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user email for which the password has to be changed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "New password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "New password confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}